fixed bug w/ removing file

+++ src/test_node.py
@@ -4,6 +4,7 @@ from node import *
 import os
 
 if __name__ == "__main__":
-    os.remove('blockchain.data')
+    if os.path.exists("blockchain.data"):
+        os.remove('blockchain.data')
     test_node = MiningNode(serialize_public_key(create_keys()[1]), int(sys.argv[1]), int(sys.argv[2]))
-    test_node.start()        
+    test_node.start()
